Bug fix for large files on cygwin

+++ src/ckoei_multidigest.c
@@ -36,7 +36,7 @@ void cko_multidigest_print(cko_multidigest_ptr x) {
   printf("\nSHA256: %s",x->hex_sha256);
   printf("\nSHA512: %s",x->hex_sha512);
   printf("\nRIPEMD160: %s",x->hex_ripemd160);
-  printf("\nSize: %lu",(unsigned long)x->size);
+  printf("\nSize: %llu",x->size);
   printf("\nVersion: %s\n",CKOEI_MULTIDIGEST_VERSION);
 }
